Configurations for entities moved to separate classes, connection string through appsettings.development.json implemented, migrations added for initial db creating

+++ Data/Entities/Product.cs
@@ -7,7 +7,7 @@ namespace EcommerceStore.Data.Entities
         public int Id { get; set; }
         public string Name { get; set; }
         public double Price { get; set; }
-        public int Quantity { get; set; }
+        public decimal Quantity { get; set; }
         public string Description { get; set; }
         public string Image { get; set; }
         public int BrandId { get; set; }
